backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary
qdrant-client==1.8.0
python-dotenv==1.0.0
langchain==0.0.339
langchain-community==0.0.29
langchain-google-genai==0.0.9
sentence-transformers==2.2.2
google-generativeai==0.3.1
python-multipart==0.0.6